Core/Loots: Fix MasterLoot & AutoLoot

+++ src/server/game/DataStores/DB2Stores.cpp
@@ -924,7 +924,7 @@ uint8 GetPowerIndexByClass(uint8 p_Class, uint8 p_Power)
         return Powers::MAX_POWERS;
 
     auto const& l_List = sPowersByClassStore[p_Class];
-    for (int l_I = 0; l_I < l_List.size(); ++l_I)
+    for (int l_I = 0; l_I < (int)l_List.size(); ++l_I)
     {
         if (l_List[l_I] == p_Power)
             return l_I;
